Module bdk::blockchain::any

Runtime-checked blockchain types

This module provides the implementation of AnyBlockchain which allows switching the inner Blockchain type at runtime.

Example

When paired with the use of ConfigurableBlockchain, it allows creating any blockchain type supported using a single line of code:

let config = serde_json::from_str("...")?;
let blockchain = AnyBlockchain::from_config(&config)?;
let height = blockchain.get_height();

Enums

AnyBlockchain

Type that can contain any of the Blockchain types defined by the library

AnyBlockchainConfig

Type that can contain any of the blockchain configurations defined by the library