Struct bdk::blockchain::compact_filters::CompactFiltersBlockchain

pub struct CompactFiltersBlockchain { /* fields omitted */ }

This is supported on feature="compact_filters" only.

Structure implementing the required blockchain traits

Example

See the blockchain::compact_filters module for a usage example.

Implementations

impl CompactFiltersBlockchain

pub fn new<P: AsRef<Path>>(
peers: Vec<Peer>,
storage_dir: P,
skip_blocks: Option<usize>
) -> Result<Self, CompactFiltersError>

This is supported on feature="compact_filters" only.

Construct a new instance given a list of peers, a path to store headers and block filters downloaded during the sync and optionally a number of blocks to ignore starting from the genesis while scanning for the wallet's outputs.

For each Peer specified a new thread will be spawned to download and verify the filters in parallel. It's currently recommended to only connect to a single peer to avoid inconsistencies in the data returned, optionally with multiple connections in parallel to speed-up the sync process.