Module bdk::blockchain::electrum

This is supported on crate feature electrum only.

Electrum

This module defines a Blockchain struct that wraps an [electrum_client::Client] and implements the logic required to populate the wallet’s database by querying the inner client.

Example

let client = electrum_client::Client::new("ssl://electrum.blockstream.info:50002")?;
let blockchain = ElectrumBlockchain::from(client);

Structs

ElectrumBlockchain

Wrapper over an Electrum Client that implements the required blockchain traits

ElectrumBlockchainConfig

Configuration for an ElectrumBlockchain