Module bdk::blockchain::esplora

This is supported on crate feature esplora only.

Esplora

This module defines a EsploraBlockchain struct that can query an Esplora backend populate the wallet’s database by:

Example

let blockchain = EsploraBlockchain::new("https://blockstream.info/testnet/api", 20);

Esplora blockchain can use either ureq or reqwest for the HTTP client depending on your needs (blocking or async respectively).

Please note, to configure the Esplora HTTP client correctly use one of: Blocking: –features=‘use-esplora-blocking’ Async: –features=‘async-interface,use-esplora-async’ –no-default-features

Structs

EsploraBlockchain

Structure that implements the logic to sync with Esplora

EsploraBlockchainConfig

Configuration for an EsploraBlockchain

Enums

EsploraError

Errors that can happen during a sync with Esplora