Struct bdk::blockchain::rpc::RpcBlockchainFactory

pub struct RpcBlockchainFactory {
    pub url: String,
    pub auth: Auth,
    pub network: Network,
    pub wallet_name_prefix: Option<String>,
    pub default_skip_blocks: u32,
    pub sync_params: Option<RpcSyncParams>,
}

Available on crate feature rpc only.

Factory of RpcBlockchain instances, implements BlockchainFactory

Internally caches the node url and authentication params and allows getting many different RpcBlockchain objects for different wallet names and with different rescan heights.

Example

let factory = RpcBlockchainFactory {
    url: "http://127.0.0.1:18332".to_string(),
    auth: Auth::Cookie {
        file: "/home/user/.bitcoin/.cookie".into(),
    },
    network: Network::Testnet,
    wallet_name_prefix: Some("prefix-".to_string()),
    default_skip_blocks: 100_000,
    sync_params: None,
};
let main_wallet_blockchain = factory.build("main_wallet", Some(200_000))?;

Fields
url: String

The bitcoin node url

auth: Auth

The bitcoin node authentication mechanism

network: Network

The network we are using (it will be checked the bitcoin node network matches this)

wallet_name_prefix: Option<String>

The optional prefix used to build the full wallet name for blockchains

default_skip_blocks: u32

Default number of blocks to skip which will be inherited by blockchain unless overridden

sync_params: Option<RpcSyncParams>

Sync parameters