Struct bdk::blockchain::EsploraBlockchain

pub struct EsploraBlockchain { /* private fields */ }

Available on crate feature esplora only.

Structure that implements the logic to sync with Esplora

Example

See the blockchain::esplora module for a usage example.

Implementations

impl EsploraBlockchain

pub fn new(base_url: &str, stop_gap: usize) -> Self

Create a new instance of the client from a base URL and the stop_gap.

pub fn from_client(url_client: BlockingClient, stop_gap: usize) -> Self

Build a new instance given a client

pub fn with_concurrency(self, concurrency: u8) -> Self

Set the number of parallel requests the client can make.