Trait bdk::blockchain::BlockchainMarker

pub trait BlockchainMarker { }

Marker trait for a blockchain backend

This is a marker trait for blockchain types. It is automatically implemented for types that implement Blockchain, so as a user of the library you won't have to implement this manually.

Users of the library will probably never have to implement this trait manually, but they could still need to import it to define types and structs with generics; Implementing only the marker trait is pointless, since OfflineBlockchain already does that, and whenever Blockchain is implemented, the marker trait is also automatically implemented by the library.