Trait bdk::blockchain::Progress

pub trait Progress: Send + 'static + Debug {
    fn update(&self, progress: f32, message: Option<String>) -> Result<(), Error>;
}

Trait for types that can receive and process progress updates during WalletSync::wallet_sync and WalletSync::wallet_setup

Required Methods

fn update(&self, progress: f32, message: Option<String>) -> Result<(), Error>

Send a new progress update

The progress value should be in the range 0.0 - 100.0, and the message value is an optional text message that can be displayed to the user.