Struct bdk::database::memory::MemoryDatabase

pub struct MemoryDatabase { /* private fields */ }

In-memory ephemeral database

This database can be used as a temporary storage for wallets that are not kept permanently on a device, or on platforms that don’t provide a filesystem, like wasm32.

Once it’s dropped its content will be lost.

If you are looking for a permanent storage solution, you can try with the default key-value database called [sled]. See the database module documentation for more details.

Implementations

impl MemoryDatabase

pub fn new() -> Self

Create a new empty database