Struct bdk::database::SqliteDatabase

pub struct SqliteDatabase {
    pub path: PathBuf,
    pub connection: Connection,
}

Sqlite database stored on filesystem

This is a permanent storage solution for devices and platforms that provide a filesystem. crate::database

Fields
path: PathBuf

Path on the local filesystem to store the sqlite file

connection: Connection

A rusqlite connection object to the sqlite database

Implementations

impl SqliteDatabase

pub fn new<T: AsRef<Path>>(path: T) -> Self

Instantiate a new SqliteDatabase instance by creating a connection to the database stored at path