Module bdk::descriptor

Descriptors

This module contains generic utilities to work with descriptors, plus some re-exported types from [miniscript].

Re-exports

pub use self::checksum::get_checksum;

pub use self::policy::Policy;

Modules

checksum
Descriptor checksum

error
Descriptor errors

policy
Descriptor policy

Structs

Miniscript
Top-level script AST type

Enums

Descriptor
Script descriptor

Legacy
Legacy ScriptContext

Segwitv0
Segwitv0 ScriptContext

Terminal
All AST elements

Traits

ExtractPolicy
Trait implemented on Descriptors to add a method to extract the spending policy

MiniscriptKey
Public key trait which can be converted to Hash type

ScriptContext

ToPublicKey
Trait describing public key types which can be converted to bitcoin pubkeys

Type Definitions

ExtendedDescriptor
Alias for a Descriptor that can contain extended keys using [DescriptorPublicKey]

HDKeyPaths
Alias for the type of maps that represent derivation paths in a psbt::Input or psbt::Output