Enum bdk::descriptor::Descriptor

pub enum Descriptor<Pk>where
    Pk: MiniscriptKey,{
    Bare(Bare<Pk>),
    Pkh(Pkh<Pk>),
    Wpkh(Wpkh<Pk>),
    Sh(Sh<Pk>),
    Wsh(Wsh<Pk>),
    Tr(Tr<Pk>),
}

Script descriptor

Variants

Bare(Bare<Pk>)

A raw scriptpubkey (including pay-to-pubkey) under Legacy context

Pkh(Pkh<Pk>)

Pay-to-PubKey-Hash

Wpkh(Wpkh<Pk>)

Pay-to-Witness-PubKey-Hash

Sh(Sh<Pk>)

Pay-to-ScriptHash(includes nested wsh/wpkh/sorted multi)

Wsh(Wsh<Pk>)

Pay-to-Witness-ScriptHash with Segwitv0 context

Tr(Tr<Pk>)

Pay-to-Taproot

Implementations

impl<Pk> Descriptor<Pk>where
Pk: MiniscriptKey,

pub fn new_pk(pk: Pk) -> Descriptor<Pk>

Create a new pk descriptor

pub fn new_pkh(pk: Pk) -> Descriptor<Pk>

Create a new PkH descriptor

pub fn new_wpkh(pk: Pk) -> Result<Descriptor<Pk>, Error>

Create a new Wpkh descriptor Will return Err if uncompressed key is used

pub fn new_sh_wpkh(pk: Pk) -> Result<Descriptor<Pk>, Error>

Create a new sh wrapped wpkh from Pk. Errors when uncompressed keys are supplied

pub fn new_sh(ms: Miniscript<Pk, Legacy>) -> Result<Descriptor<Pk>, Error>

Create a new sh for a given redeem script Errors when miniscript exceeds resource limits under p2sh context or does not type check at the top level

pub fn new_wsh(ms: Miniscript<Pk, Segwitv0>) -> Result<Descriptor<Pk>, Error>

Create a new wsh descriptor from witness script Errors when miniscript exceeds resource limits under p2sh context or does not type check at the top level

pub fn new_sh_wsh(ms: Miniscript<Pk, Segwitv0>) -> Result<Descriptor<Pk>, Error>

Create a new sh wrapped wsh descriptor with witness script Errors when miniscript exceeds resource limits under wsh context or does not type check at the top level

pub fn new_bare(ms: Miniscript<Pk, BareCtx>) -> Result<Descriptor<Pk>, Error>

Create a new bare descriptor from witness script Errors when miniscript exceeds resource limits under bare context or does not type check at the top level

pub fn new_sh_with_wpkh(wpkh: Wpkh<Pk>) -> Descriptor<Pk>

Create a new sh wrapper for the given wpkh descriptor

pub fn new_sh_with_wsh(wsh: Wsh<Pk>) -> Descriptor<Pk>

Create a new sh wrapper for the given wsh descriptor

pub fn new_sh_sortedmulti(
k: usize,
pks: Vec<Pk, Global>
) -> Result<Descriptor<Pk>, Error>

Create a new sh sortedmulti descriptor with threshold k and Vec of pks. Errors when miniscript exceeds resource limits under p2sh context

pub fn new_sh_wsh_sortedmulti(
k: usize,
pks: Vec<Pk, Global>
) -> Result<Descriptor<Pk>, Error>

Create a new sh wrapped wsh sortedmulti descriptor from threshold k and Vec of pks Errors when miniscript exceeds resource limits under segwit context

pub fn new_wsh_sortedmulti(
k: usize,
pks: Vec<Pk, Global>
) -> Result<Descriptor<Pk>, Error>

Create a new wsh sorted multi descriptor Errors when miniscript exceeds resource limits under p2sh context

pub fn new_tr(
key: Pk,
script: Option<TapTree<Pk>>
) -> Result<Descriptor<Pk>, Error>

Create new tr descriptor Errors when miniscript exceeds resource limits under Tap context

pub fn desc_type(&self) -> DescriptorType

Get the [DescriptorType] of Descriptor

pub fn sanity_check(&self) -> Result<(), Error>

Checks whether the descriptor is safe.

Checks whether all the spend paths in the descriptor are possible on the bitcoin network under the current standardness and consensus rules. Also checks whether the descriptor requires signatures on all spend paths and whether the script is malleable.

In general, all the guarantees of miniscript hold only for safe scripts. The signer may not be able to find satisfactions even if one exists.

pub fn max_satisfaction_weight(&self) -> Result<usize, Error>

Computes an upper bound on the weight of a satisfying witness to the transaction.

Assumes all ec-signatures are 73 bytes, including push opcode and sighash suffix. Includes the weight of the VarInts encoding the scriptSig and witness stack length.

Errors

When the descriptor is impossible to safisfy (ex: sh(OP_FALSE)).

impl<Pk> Descriptor<Pk>where
Pk: MiniscriptKey + ToPublicKey,

pub fn address(&self, network: Network) -> Result<Address, Error>

Computes the Bitcoin address of the descriptor, if one exists

Some descriptors like pk() don’t have an address.

Errors

For raw/bare descriptors that don’t have an address.

pub fn script_pubkey(&self) -> Script

Computes the scriptpubkey of the descriptor.

pub fn unsigned_script_sig(&self) -> Script

Computes the scriptSig that will be in place for an unsigned input spending an output with this descriptor. For pre-segwit descriptors, which use the scriptSig for signatures, this returns the empty script.

This is used in Segwit transactions to produce an unsigned transaction whose txid will not change during signing (since only the witness data will change).

pub fn explicit_script(&self) -> Result<Script, Error>

Computes the the underlying script before any hashing is done. For Bare, Pkh and Wpkh this is the scriptPubkey; for ShWpkh and Sh this is the redeemScript; for the others it is the witness script.

Errors

If the descriptor is a taproot descriptor.

pub fn script_code(&self) -> Result<Script, Error>

Computes the scriptCode of a transaction output.

The scriptCode is the Script of the previous transaction output being serialized in the sighash when evaluating a CHECKSIG & co. OP code.

Errors

If the descriptor is a taproot descriptor.

pub fn get_satisfaction<S>(
&self,
satisfier: S
) -> Result<(Vec<Vec<u8, Global>, Global>, Script), Error>where
S: Satisfier<Pk>,

Returns satisfying non-malleable witness and scriptSig to spend an output controlled by the given descriptor if it possible to construct one using the satisfier S.

pub fn get_satisfaction_mall<S>(
&self,
satisfier: S
) -> Result<(Vec<Vec<u8, Global>, Global>, Script), Error>where
S: Satisfier<Pk>,

Returns a possilbly mallable satisfying non-malleable witness and scriptSig to spend an output controlled by the given descriptor if it possible to construct one using the satisfier S.

pub fn satisfy<S>(&self, txin: &mut TxIn, satisfier: S) -> Result<(), Error>where
S: Satisfier<Pk>,

Attempts to produce a non-malleable satisfying witness and scriptSig to spend an output controlled by the given descriptor; add the data to a given TxIn output.

impl Descriptor<DescriptorPublicKey>

pub fn is_deriveable(&self) -> bool

👎Deprecated: use has_wildcards instead

Whether or not the descriptor has any wildcards

pub fn has_wildcard(&self) -> bool

Whether or not the descriptor has any wildcards i.e. /*.

pub fn at_derivation_index(
&self,
index: u32
) -> Descriptor<DefiniteDescriptorKey>

Replaces all wildcards (i.e. /*) in the descriptor with a particular derivation index, turning it into a definite descriptor.

Panics

If index ≥ 2^31

pub fn derive(&self, index: u32) -> Descriptor<DefiniteDescriptorKey>

👎Deprecated: use at_derivation_index instead

Deprecated name for [at_derivation_index].

pub fn derived_descriptor<C>(
&self,
secp: &Secp256k1<C>,
index: u32
) -> Result<Descriptor<PublicKey>, ConversionError>where
C: Verification,

Convert all the public keys in the descriptor to [bitcoin::PublicKey] by deriving them or otherwise converting them. All [bitcoin::XOnlyPublicKey]s are converted to by adding a default(0x02) y-coordinate.

This is a shorthand for:

    .expect("Valid ranged descriptor");
let derived_descriptor = descriptor.at_derivation_index(index).derived_descriptor(&secp);

and is only here really here for backwards compatbility. See at_derivation_index and [derived_descriptor] for more documentation.

Errors

This function will return an error if hardened derivation is attempted.

pub fn parse_descriptor<C>(
secp: &Secp256k1<C>,
s: &str
) -> Result<(Descriptor<DescriptorPublicKey>, HashMap<DescriptorPublicKey, DescriptorSecretKey, RandomState>), Error>where
C: Signing,

Parse a descriptor that may contain secret keys

Internally turns every secret key found into the corresponding public key and then returns a a descriptor that only contains public keys and a map to lookup the secret key given a public key.

pub fn to_string_with_secret(
&self,
key_map: &HashMap<DescriptorPublicKey, DescriptorSecretKey, RandomState>
) -> String

Serialize a descriptor to string with its secret keys

pub fn find_derivation_index_for_spk<C>(
&self,
secp: &Secp256k1<C>,
script_pubkey: &Script,
range: Range<u32>
) -> Result<Option<(u32, Descriptor<PublicKey>)>, ConversionError>where
C: Verification,

Utility method for deriving the descriptor at each index in a range to find one matching script_pubkey.

If it finds a match then it returns the index it was derived at and the concrete descriptor at that index. If the descriptor is non-derivable then it will simply check the script pubkey against the descriptor and return it if it matches (in this case the index returned will be meaningless).

impl Descriptor<DefiniteDescriptorKey>

pub fn derived_descriptor<C>(
&self,
secp: &Secp256k1<C>
) -> Result<Descriptor<PublicKey>, ConversionError>where
C: Verification,

Convert all the public keys in the descriptor to [bitcoin::PublicKey] by deriving them or otherwise converting them. All [bitcoin::XOnlyPublicKey]s are converted to by adding a default(0x02) y-coordinate.

Examples

use miniscript::descriptor::{Descriptor, DescriptorPublicKey};
use miniscript::bitcoin::secp256k1;
use std::str::FromStr;

// test from bip 86
let secp = secp256k1::Secp256k1::verification_only();
let descriptor = Descriptor::<DescriptorPublicKey>::from_str("tr(xpub6BgBgsespWvERF3LHQu6CnqdvfEvtMcQjYrcRzx53QJjSxarj2afYWcLteoGVky7D3UKDP9QyrLprQ3VCECoY49yfdDEHGCtMMj92pReUsQ/0/*)")
    .expect("Valid ranged descriptor");
let result = descriptor.at_derivation_index(0).derived_descriptor(&secp).expect("Non-hardened derivation");
assert_eq!(result.to_string(), "tr(03cc8a4bc64d897bddc5fbc2f670f7a8ba0b386779106cf1223c6fc5d7cd6fc115)#6qm9h8ym");

Errors

This function will return an error if hardened derivation is attempted.