Trait bdk::descriptor::ToWalletDescriptor

pub trait ToWalletDescriptor {
    pub fn to_wallet_descriptor(
        self, 
        network: Network
    ) -> Result<(ExtendedDescriptor, KeyMap), KeyError>;
}

Trait for types which can be converted into an ExtendedDescriptor and a KeyMap usable by a wallet in a specific [Network]

Required methods

pub fn to_wallet_descriptor(
self,
network: Network
) -> Result<(ExtendedDescriptor, KeyMap), KeyError>

Convert to wallet descriptor

Loading content...