Module bdk::keys::bip39

This is supported on crate feature keys-bip39 only.

BIP-0039

Structs

Mnemonic
The primary type in this crate, most tasks require creating or using one.

Seed
The secret value used to derive HD wallet addresses from a Mnemonic phrase.

Enums

Language
The language determines which words will be used in a mnemonic phrase, but also indirectly determines the binary value of each word when a Mnemonic is turned into a Seed.

MnemonicType
Determines the number of words that will be present in a Mnemonic phrase

Type Definitions

MnemonicWithPassphrase
Type for a BIP39 mnemonic with an optional passphrase