Enum bdk::keys::bip39::Language

pub enum Language {
    English,
    ChineseSimplified,
    ChineseTraditional,
    French,
    Italian,
    Japanese,
    Korean,
    Spanish,
}

This is supported on crate feature keys-bip39 only.

The language determines which words will be used in a mnemonic phrase, but also indirectly determines the binary value of each word when a Mnemonic is turned into a Seed.

These are not of much use right now, and may even be removed from the crate, as there is no official language specified by the standard except English.

Variants

English

ChineseSimplified

ChineseTraditional

French

Italian

Japanese

Korean

Spanish

Implementations

impl Language

pub fn from_language_code(language_code: &str) -> Option<Language>

Construct a word list from its language code. Returns None if the language code is not valid or not supported.

pub fn wordlist(&self) -> &'static WordList

Get the word list for this language

pub fn wordmap(&self) -> &'static WordMap

Get a [WordMap][WordMap] that allows word -> index lookups in the word list

The index of an individual word in the word list is used as the binary value of that word when the phrase is turned into a [Seed][Seed].