Module bdk::wallet

Wallet

This module defines the Wallet structure.

Modules

coin_selection

Coin selection

export

Wallet export

hardwaresignerhardware-signer

HWI Signer

signer

Generalized signers

time

Cross-platform time

tx_builder

Transaction builder

verifyverify

Verify transactions against the consensus rules

Structs

AddressInfo

A derived address and the index it was found at. For convenience this automatically derefs to Address

SyncOptions

Options to a sync.

Wallet

A Bitcoin wallet

Enums

AddressIndex

The address index selection strategy to use to derived an address from the wallet’s external descriptor. See Wallet::get_address. If you’re unsure which one to use use WalletIndex::New.

Traits

IsDust

Trait to check if a value is below the dust limit. We are performing dust value calculation for a given script public key using rust-bitcoin to keep it compatible with network dust rate

Functions

get_funded_wallet

Return a fake wallet that appears to be funded for testing.

wallet_name_from_descriptor

Deterministically generate a unique name given the descriptors defining the wallet