Trait bdk::wallet::address_validator::AddressValidator

pub trait AddressValidator: Send + Sync + Debug {
    fn validate(
        &self, 
        keychain: KeychainKind, 
        hd_keypaths: &HDKeyPaths, 
        script: &Script
    ) -> Result<(), AddressValidatorError>;
}

Trait to build address validators

All the address validators attached to a wallet with Wallet::add_address_validator will be polled every time an address (external or internal) is generated by the wallet. Errors returned in the validator will be propagated up to the original caller that triggered the address generation.

For a usage example see this module’s documentation.

Required methods

fn validate(
&self,
keychain: KeychainKind,
hd_keypaths: &HDKeyPaths,
script: &Script
) -> Result<(), AddressValidatorError>

Validate or inspect an address

Loading content...