Enum bdk::wallet::signer::SignerError

pub enum SignerError {

    MissingKey,
    InvalidKey,
    UserCanceled,
    InputIndexOutOfRange,
    MissingNonWitnessUtxo,
    InvalidNonWitnessUtxo,
    MissingWitnessUtxo,
    MissingWitnessScript,
    MissingHdKeypath,
    NonStandardSighash,
    InvalidSighash,
    SighashError(Error),
    HWIError(Error),

}

Signing error

Variants

MissingKey

The private key is missing for the required public key

InvalidKey

The private key in use has the right fingerprint but derives differently than expected

UserCanceled

The user canceled the operation

InputIndexOutOfRange

Input index is out of range

MissingNonWitnessUtxo

The non_witness_utxo field of the transaction is required to sign this input

InvalidNonWitnessUtxo

The non_witness_utxo specified is invalid

MissingWitnessUtxo

The witness_utxo field of the transaction is required to sign this input

MissingWitnessScript

The witness_script field of the transaction is required to sign this input

MissingHdKeypath

The fingerprint and derivation path are missing from the psbt input

NonStandardSighash

The psbt contains a non-SIGHASH_ALL sighash in one of its input and the user hasnโt explicitly allowed them

To enable signing transactions with non-standard sighashes set SignOptions::allow_all_sighashes to true.

InvalidSighash

Invalid SIGHASH for the signing context in use

SighashError(Error)

Error while computing the hash to sign

HWIError(Error)

Error while signing using hardware wallets