Trait bdk::wallet::signer::TransactionSigner

pub trait TransactionSigner: SignerCommon {
    fn sign_transaction(
        &self,
        psbt: &mut PartiallySignedTransaction,
        sign_options: &SignOptions,
        secp: &Secp256k1<All>
    ) -> Result<(), SignerError>;
}

PSBT signer

This trait can be implemented when the signer can’t sign inputs individually, but signs the whole transaction at once.

Required Methods

fn sign_transaction(
&self,
psbt: &mut PartiallySignedTransaction,
sign_options: &SignOptions,
secp: &Secp256k1<All>
) -> Result<(), SignerError>

Sign all the inputs of the psbt