Struct bdk::wallet::AddressInfo

pub struct AddressInfo {
    pub index: u32,
    pub address: Address,
}

A derived address and the index it was found at For convenience this automatically derefs to Address

Fields
index: u32

Child index of this address

address: Address

Address