Module bdk::blockchain::esplora

This is supported on crate feature esplora only.

Esplora

This module defines a Blockchain struct that can query an Esplora backend populate the wallet's database by

Example

let blockchain = EsploraBlockchain::new("https://blockstream.info/testnet/api", None);

Structs

EsploraBlockchain
Structure that implements the logic to sync with Esplora

EsploraBlockchainConfig
Configuration for an EsploraBlockchain

Enums

EsploraError
Errors that can happen during a sync with EsploraBlockchain