Trait bdk::blockchain::Blockchain

pub trait Blockchain: WalletSync + GetHeight + GetTx + GetBlockHash {
    fn get_capabilities(&self) -> HashSet<Capability>;
    fn broadcast(&self, tx: &Transaction) -> Result<(), Error>;
    fn estimate_fee(&self, target: usize) -> Result<FeeRate, Error>;
}

Trait that defines the actions that must be supported by a blockchain backend

Required Methods

fn get_capabilities(&self) -> HashSet<Capability>

Return the set of Capability supported by this backend

fn broadcast(&self, tx: &Transaction) -> Result<(), Error>

Broadcast a transaction

fn estimate_fee(&self, target: usize) -> Result<FeeRate, Error>

Estimate the fee rate required to confirm a transaction in a given target of blocks