Trait bdk::blockchain::Progress

pub trait Progress: Send {
    fn update(
        &self, 
        progress: f32, 
        message: Option<String>
    ) -> Result<(), Error>;
}

Trait for types that can receive and process progress updates during Blockchain::sync and Blockchain::setup

Required methods

fn update(&self, progress: f32, message: Option<String>) -> Result<(), Error>

Send a new progress update

The progress value should be in the range 0.0 - 100.0, and the message value is an optional text message that can be displayed to the user.