Trait bdk::database::BatchOperations

pub trait BatchOperations {

    fn set_script_pubkey(
        &mut self, 
        script: &Script, 
        keychain: KeychainKind, 
        child: u32
    ) -> Result<(), Error>;

    fn set_utxo(&mut self, utxo: &LocalUtxo) -> Result<(), Error>;

    fn set_raw_tx(&mut self, transaction: &Transaction) -> Result<(), Error>;

    fn set_tx(&mut self, transaction: &TransactionDetails) -> Result<(), Error>;

    fn set_last_index(
        &mut self, 
        keychain: KeychainKind, 
        value: u32
    ) -> Result<(), Error>;

    fn set_sync_time(&mut self, sync_time: SyncTime) -> Result<(), Error>;

    fn del_script_pubkey_from_path(
        &mut self, 
        keychain: KeychainKind, 
        child: u32
    ) -> Result<Option<Script>, Error>;

    fn del_path_from_script_pubkey(
        &mut self, 
        script: &Script
    ) -> Result<Option<(KeychainKind, u32)>, Error>;

    fn del_utxo(
        &mut self, 
        outpoint: &OutPoint
    ) -> Result<Option<LocalUtxo>, Error>;

    fn del_raw_tx(&mut self, txid: &Txid) -> Result<Option<Transaction>, Error>;

    fn del_tx(
        &mut self, 
        txid: &Txid, 
        include_raw: bool
    ) -> Result<Option<TransactionDetails>, Error>;

    fn del_last_index(
        &mut self, 
        keychain: KeychainKind
    ) -> Result<Option<u32>, Error>;

    fn del_sync_time(&mut self) -> Result<Option<SyncTime>, Error>;

}

Trait for operations that can be batched

This trait defines the list of operations that must be implemented on the Database type and the BatchDatabase::Batch type.

Required methods

fn set_script_pubkey(
&mut self,
script: &Script,
keychain: KeychainKind,
child: u32
) -> Result<(), Error>

Store a script_pubkey along with its keychain and child number.

fn set_utxo(&mut self, utxo: &LocalUtxo) -> Result<(), Error>

Store a LocalUtxo

fn set_raw_tx(&mut self, transaction: &Transaction) -> Result<(), Error>

Store a raw transaction

fn set_tx(&mut self, transaction: &TransactionDetails) -> Result<(), Error>

Store the metadata of a transaction

fn set_last_index(
&mut self,
keychain: KeychainKind,
value: u32
) -> Result<(), Error>

Store the last derivation index for a given keychain.

fn set_sync_time(&mut self, sync_time: SyncTime) -> Result<(), Error>

Store the sync time

fn del_script_pubkey_from_path(
&mut self,
keychain: KeychainKind,
child: u32
) -> Result<Option<Script>, Error>

Delete a script_pubkey given the keychain and its child number.

fn del_path_from_script_pubkey(
&mut self,
script: &Script
) -> Result<Option<(KeychainKind, u32)>, Error>

Delete the data related to a specific script_pubkey, meaning the keychain and the child number.

fn del_utxo(&mut self, outpoint: &OutPoint) -> Result<Option<LocalUtxo>, Error>

Delete a LocalUtxo given its [OutPoint]

fn del_raw_tx(&mut self, txid: &Txid) -> Result<Option<Transaction>, Error>

Delete a raw transaction given its [Txid]

fn del_tx(
&mut self,
txid: &Txid,
include_raw: bool
) -> Result<Option<TransactionDetails>, Error>

Delete the metadata of a transaction and optionally the raw transaction itself

fn del_last_index(
&mut self,
keychain: KeychainKind
) -> Result<Option<u32>, Error>

Delete the last derivation index for a keychain.

fn del_sync_time(&mut self) -> Result<Option<SyncTime>, Error>

Reset the sync time to None

Returns the removed value