Enum bdk::descriptor::Descriptor

pub enum Descriptor<Pk> where
    Pk: MiniscriptKey,  {
    Bare(Miniscript<Pk, Bare>),
    Pk(Pk),
    Pkh(Pk),
    Wpkh(Pk),
    ShWpkh(Pk),
    Sh(Miniscript<Pk, Legacy>),
    Wsh(Miniscript<Pk, Segwitv0>),
    ShWsh(Miniscript<Pk, Segwitv0>),
    ShSortedMulti(SortedMultiVec<Pk, Legacy>),
    WshSortedMulti(SortedMultiVec<Pk, Segwitv0>),
    ShWshSortedMulti(SortedMultiVec<Pk, Segwitv0>),
}

Script descriptor

Variants

Bare(Miniscript<Pk, Bare>)

A raw scriptpubkey (including pay-to-pubkey) under Legacy context

Pk(Pk)

Pay-to-Pubkey

Pkh(Pk)

Pay-to-PubKey-Hash

Wpkh(Pk)

Pay-to-Witness-PubKey-Hash

ShWpkh(Pk)

Pay-to-Witness-PubKey-Hash inside P2SH

Sh(Miniscript<Pk, Legacy>)

Pay-to-ScriptHash with Legacy context

Wsh(Miniscript<Pk, Segwitv0>)

Pay-to-Witness-ScriptHash with Segwitv0 context

ShWsh(Miniscript<Pk, Segwitv0>)

P2SH-P2WSH with Segwitv0 context

ShSortedMulti(SortedMultiVec<Pk, Legacy>)

Sortedmulti under P2SH

WshSortedMulti(SortedMultiVec<Pk, Segwitv0>)

Sortedmulti under P2WSH

ShWshSortedMulti(SortedMultiVec<Pk, Segwitv0>)

Sortedmulti under P2SH-P2WSH

Implementations

impl<Pk> Descriptor<Pk> where
Pk: MiniscriptKey,

pub fn translate_pk<Fpk, Fpkh, Q, E>(
&self,
translatefpk: Fpk,
translatefpkh: Fpkh
) -> Result<Descriptor<Q>, E> where
Fpk: FnMut(&Pk) -> Result<Q, E>,
Fpkh: FnMut(&<Pk as MiniscriptKey>::Hash) -> Result<<Q as MiniscriptKey>::Hash, E>,
Q: MiniscriptKey,

Convert a descriptor using abstract keys to one using specific keys This will panic if translatefpk returns an uncompressed key when converting to a Segwit descriptor. To prevent this panic, ensure translatefpk returns an error in this case instead.

pub fn sanity_check(&self) -> Result<(), Error>

Whether the descriptor is safe Checks whether all the spend paths in the descriptor are possible on the bitcoin network under the current standardness and consensus rules Also checks whether the descriptor requires signauture on all spend paths And whether the script is malleable. In general, all the guarantees of miniscript hold only for safe scripts. All the analysis guarantees of miniscript only hold safe scripts. The signer may not be able to find satisfactions even if one exists

impl<Pk> Descriptor<Pk> where
Pk: MiniscriptKey,

pub fn address<ToPkCtx>(
&self,
network: Network,
to_pk_ctx: ToPkCtx
) -> Option<Address> where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Computes the Bitcoin address of the descriptor, if one exists to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait [trait.ToPublicKey]

pub fn script_pubkey<ToPkCtx>(&self, to_pk_ctx: ToPkCtx) -> Script where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Computes the scriptpubkey of the descriptor to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait ToPublicKey

pub fn unsigned_script_sig<ToPkCtx>(&self, to_pk_ctx: ToPkCtx) -> Script where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Computes the scriptSig that will be in place for an unsigned input spending an output with this descriptor. For pre-segwit descriptors, which use the scriptSig for signatures, this returns the empty script.

This is used in Segwit transactions to produce an unsigned transaction whose txid will not change during signing (since only the witness data will change). to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait ToPublicKey

pub fn witness_script<ToPkCtx>(&self, to_pk_ctx: ToPkCtx) -> Script where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Computes the "witness script" of the descriptor, i.e. the underlying script before any hashing is done. For Bare, Pkh and Wpkh this is the scriptPubkey; for ShWpkh and Sh this is the redeemScript; for the others it is the witness script. to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait ToPublicKey

pub fn get_satisfication<ToPkCtx, S>(
&self,
satisfier: S,
to_pk_ctx: ToPkCtx
) -> Result<(Vec<Vec<u8, Global>, Global>, Script), Error> where
Pk: ToPublicKey<ToPkCtx>,
S: Satisfier<ToPkCtx, Pk>,
ToPkCtx: Copy,

Returns satisfying witness and scriptSig to spend an output controlled by the given descriptor if it possible to construct one using the satisfier S. to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait ToPublicKey

pub fn satisfy<ToPkCtx, S>(
&self,
txin: &mut TxIn,
satisfier: S,
to_pk_ctx: ToPkCtx
) -> Result<(), Error> where
Pk: ToPublicKey<ToPkCtx>,
S: Satisfier<ToPkCtx, Pk>,
ToPkCtx: Copy,

Attempts to produce a satisfying witness and scriptSig to spend an output controlled by the given descriptor; add the data to a given TxIn output.

pub fn max_satisfaction_weight<ToPkCtx>(
&self,
to_pk_ctx: ToPkCtx
) -> Option<usize> where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Computes an upper bound on the weight of a satisfying witness to the transaction. Assumes all signatures are 73 bytes, including push opcode and sighash suffix. Includes the weight of the VarInts encoding the scriptSig and witness stack length.

pub fn script_code<ToPkCtx>(&self, to_pk_ctx: ToPkCtx) -> Script where
Pk: ToPublicKey<ToPkCtx>,
ToPkCtx: Copy,

Get the scriptCode of a transaction output.

The scriptCode is the Script of the previous transaction output being serialized in the sighash when evaluating a CHECKSIG & co. OP code. to_pk_ctx denotes the ToPkCtx required for deriving bitcoin::PublicKey from MiniscriptKey using ToPublicKey. If MiniscriptKey is already is [bitcoin::PublicKey], then the context would be [NullCtx] and [descriptor.DescriptorPublicKeyCtx] if MiniscriptKey is [descriptor.DescriptorPublicKey]

In general, this is defined by generic for the trait ToPublicKey

impl Descriptor<DescriptorPublicKey>

pub fn derive(
&self,
child_number: ChildNumber
) -> Descriptor<DescriptorPublicKey>

Derives all wildcard keys in the descriptor using the supplied child_number

pub fn parse_descriptor(
s: &str
) -> Result<(Descriptor<DescriptorPublicKey>, HashMap<DescriptorPublicKey, DescriptorSecretKey, RandomState>), Error>

Parse a descriptor that may contain secret keys

Internally turns every secret key found into the corresponding public key and then returns a a descriptor that only contains public keys and a map to lookup the secret key given a public key.

pub fn to_string_with_secret(
&self,
key_map: &HashMap<DescriptorPublicKey, DescriptorSecretKey, RandomState>
) -> String

Serialize a descriptor to string with its secret keys