Struct bdk::descriptor::Miniscript

pub struct Miniscript<Pk, Ctx> where
    Ctx: ScriptContext,
    Pk: MiniscriptKey,  {
    pub node: Terminal<Pk, Ctx>,
    pub ty: Type,
    pub ext: ExtData,
    // some fields omitted
}

Top-level script AST type

Fields
node: Terminal<Pk, Ctx>

A node in the Abstract Syntax Tree(

ty: Type

The correctness and malleability type information for the AST node

ext: ExtData

Additional information helpful for extra analysis.

Implementations

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn requires_sig(&self) -> bool

Whether all spend paths of miniscript require a signature

pub fn is_non_malleable(&self) -> bool

Whether the miniscript is malleable

pub fn within_resource_limits(&self) -> bool

Whether the miniscript can exceed the resource limits(Opcodes, Stack limit etc)

pub fn has_mixed_timelocks(&self) -> bool

Whether the miniscript contains a combination of timelocks

pub fn has_repeated_keys(&self) -> bool

Whether the miniscript has repeated Pk or Pkh

pub fn sanity_check(&self) -> Result<(), AnalysisError>

Check whether the underlying Miniscript is safe under the current context Lifting these polices would create a semantic representation that does not represent the underlying semantics when miniscript is spent. Signing logic may not find satisfaction even if one exists.

For most cases, users should be dealing with safe scripts. Use this function to check whether the guarantees of library hold. Most functions of the library like would still work, but results cannot be relied upon

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

Iterator-related extensions for Miniscript

pub fn iter(&self) -> Iter<'_, Pk, Ctx>

Creates a new [Iter] iterator that will iterate over all Miniscript items within AST by traversing its branches. For the specific algorithm please see [Iter::next] function.

pub fn iter_pk(&self) -> PkIter<'_, Pk, Ctx>

Creates a new [PkIter] iterator that will iterate over all plain public keys (and not key hash values) present in Miniscript items within AST by traversing all its branches. For the specific algorithm please see [PkIter::next] function.

pub fn iter_pkh(&self) -> PkhIter<'_, Pk, Ctx>

Creates a new [PkhIter] iterator that will iterate over all public keys hashes (and not plain public keys) present in Miniscript items within AST by traversing all its branches. For the specific algorithm please see [PkhIter::next] function.

pub fn iter_pk_pkh(&self) -> PkPkhIter<'_, Pk, Ctx>

Creates a new [PkPkhIter] iterator that will iterate over all plain public keys and key hash values present in Miniscript items within AST by traversing all its branches. For the specific algorithm please see [PkPkhIter::next] function.

pub fn branches(&self) -> Vec<&Miniscript<Pk, Ctx>, Global>

Enumerates all child nodes of the current AST node (self) and returns a Vec referencing them.

pub fn get_nth_child(&self, n: usize) -> Option<&Miniscript<Pk, Ctx>>

Returns child node with given index, if any

pub fn get_leaf_pk(&self) -> Vec<Pk, Global>

Returns Vec with cloned version of all public keys from the current miniscript item, if any. Otherwise returns an empty Vec.

NB: The function analyzes only single miniscript item and not any of its descendants in AST. To obtain a list of all public keys within AST use Miniscript::iter_pk() function, for example miniscript.iter_pubkeys().collect().

pub fn get_leaf_pkh(&self) -> Vec<<Pk as MiniscriptKey>::Hash, Global>

Returns Vec with hashes of all public keys from the current miniscript item, if any. Otherwise returns an empty Vec.

For each public key the function computes hash; for each hash of the public key the function returns its cloned copy.

NB: The function analyzes only single miniscript item and not any of its descendants in AST. To obtain a list of all public key hashes within AST use Miniscript::iter_pkh() function, for example miniscript.iter_pubkey_hashes().collect().

pub fn get_leaf_pk_pkh(&self) -> Vec<PkPkh<Pk>, Global>

Returns Vec of [PkPkh] entries, representing either public keys or public key hashes, depending on the data from the current miniscript item. If there is no public keys or hashes, the function returns an empty Vec.

NB: The function analyzes only single miniscript item and not any of its descendants in AST. To obtain a list of all public keys or hashes within AST use Miniscript::iter_pk_pkh() function, for example miniscript.iter_pubkeys_and_hashes().collect().

pub fn get_nth_pk(&self, n: usize) -> Option<Pk>

Returns Option::Some with cloned n'th public key from the current miniscript item, if any. Otherwise returns Option::None.

NB: The function analyzes only single miniscript item and not any of its descendants in AST.

pub fn get_nth_pkh(&self, n: usize) -> Option<<Pk as MiniscriptKey>::Hash>

Returns Option::Some with hash of n'th public key from the current miniscript item, if any. Otherwise returns Option::None.

For each public key the function computes hash; for each hash of the public key the function returns it cloned copy.

NB: The function analyzes only single miniscript item and not any of its descendants in AST.

pub fn get_nth_pk_pkh(&self, n: usize) -> Option<PkPkh<Pk>>

Returns Option::Some with hash of n'th public key or hash from the current miniscript item, if any. Otherwise returns Option::None.

NB: The function analyzes only single miniscript item and not any of its descendants in AST.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn from_ast(t: Terminal<Pk, Ctx>) -> Result<Miniscript<Pk, Ctx>, Error>

Add type information(Type and Extdata) to Miniscript based on AstElem fragment. Dependent on display and clone because of Error Display code of type_check.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn into_inner(self) -> Terminal<Pk, Ctx>

Extracts the AstElem representing the root of the miniscript

pub fn as_inner(&self) -> &Terminal<Pk, Ctx>

Get a reference to the inner AstElem representing the root of miniscript

impl<Ctx> Miniscript<PublicKey, Ctx> where
Ctx: ScriptContext,

pub fn parse_insane(
script: &Script
) -> Result<Miniscript<PublicKey, Ctx>, Error>

Attempt to parse an insane(scripts don't clear sanity checks) script into a Miniscript representation. Use this to parse scripts with repeated pubkeys, timelock mixing, malleable scripts without sig or scripts that can exceed resource limits. Some of the analysis guarantees of miniscript are lost when dealing with insane scripts. In general, in a multi-party setting users should only accept sane scripts.

pub fn parse(script: &Script) -> Result<Miniscript<PublicKey, Ctx>, Error>

Attempt to parse a Script into Miniscript representation. This function will fail parsing for scripts that do not clear the Miniscript::sanity_check checks. Use Miniscript::parse_insane to parse such scripts.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn encode(&self) -> Script where
Pk: ToPublicKey,

Encode as a Bitcoin script

pub fn script_size(&self) -> usize

Size, in bytes of the script-pubkey. If this Miniscript is used outside of segwit (e.g. in a bare or P2SH descriptor), this quantity should be multiplied by 4 to compute the weight.

In general, it is not recommended to use this function directly, but to instead call the corresponding function on a Descriptor, which will handle the segwit/non-segwit technicalities for you.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn max_satisfaction_witness_elements(&self) -> Result<usize, Error>

Maximum number of witness elements used to satisfy the Miniscript fragment, including the witness script itself. Used to estimate the weight of the VarInt that specifies this number in a serialized transaction.

This function may returns Error when the Miniscript is impossible to satisfy

pub fn max_satisfaction_size(&self) -> Result<usize, Error>

Maximum size, in bytes, of a satisfying witness. For Segwit outputs one_cost should be set to 2, since the number 1 requires two bytes to encode. For non-segwit outputs one_cost should be set to 1, since OP_1 is available in scriptSigs.

In general, it is not recommended to use this function directly, but to instead call the corresponding function on a Descriptor, which will handle the segwit/non-segwit technicalities for you.

All signatures are assumed to be 73 bytes in size, including the length prefix (segwit) or push opcode (pre-segwit) and sighash postfix.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn from_str_insane(s: &str) -> Result<Miniscript<Pk, Ctx>, Error> where
Pk: FromStr,
<Pk as MiniscriptKey>::Hash: FromStr,
<Pk as FromStr>::Err: ToString,
<<Pk as MiniscriptKey>::Hash as FromStr>::Err: ToString,

Attempt to parse an insane(scripts don't clear sanity checks) from string into a Miniscript representation. Use this to parse scripts with repeated pubkeys, timelock mixing, malleable scripts without sig or scripts that can exceed resource limits. Some of the analysis guarantees of miniscript are lost when dealing with insane scripts. In general, in a multi-party setting users should only accept sane scripts.

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn satisfy<S>(
&self,
satisfier: S
) -> Result<Vec<Vec<u8, Global>, Global>, Error> where
Pk: ToPublicKey,
S: Satisfier<Pk>,

Attempt to produce non-malleable satisfying witness for the witness script represented by the parse tree

pub fn satisfy_malleable<S>(
&self,
satisfier: S
) -> Result<Vec<Vec<u8, Global>, Global>, Error> where
Pk: ToPublicKey,
S: Satisfier<Pk>,

Attempt to produce a malleable satisfying witness for the witness script represented by the parse tree

impl<Pk, Ctx> Miniscript<Pk, Ctx> where
Ctx: ScriptContext,
Pk: MiniscriptKey,

pub fn lift_check(&self) -> Result<(), LiftError>

Lifting corresponds conversion of miniscript into Policy [policy.semantic.Policy] for human readable or machine analysis. However, naively lifting miniscripts can result in incorrect interpretations that don't correspond underlying semantics when we try to spend them on bitcoin network. This can occur if the miniscript contains a

Timelock combination

Contains a spend that exceeds resource limits