Module bdk::descriptor::template

Descriptor templates

This module contains the definition of various common script templates that are ready to be used. See the documentation of each template for an example.

Structs

BIP44
BIP44 template. Expands to pkh(key/44'/0'/0'/{0,1}/*)

BIP44Public
BIP44 public template. Expands to pkh(key/{0,1}/*)

BIP49
BIP49 template. Expands to sh(wpkh(key/49'/0'/0'/{0,1}/*))

BIP49Public
BIP49 public template. Expands to sh(wpkh(key/{0,1}/*))

BIP84
BIP84 template. Expands to wpkh(key/84'/0'/0'/{0,1}/*)

BIP84Public
BIP84 public template. Expands to wpkh(key/{0,1}/*)

P2PKH
P2PKH template. Expands to a descriptor pkh(key)

P2WPKH
P2WPKH template. Expands to a descriptor wpkh(key)

P2WPKH_P2SH
P2WPKH-P2SH template. Expands to a descriptor sh(wpkh(key))

Traits

DescriptorTemplate
Trait for descriptor templates that can be built into a full descriptor

Type Definitions

DescriptorTemplateOut
Type alias for the return type of DescriptorTemplate, descriptor! and others