Trait bdk::descriptor::ScriptContext

pub trait ScriptContext: Debug + Clone + Ord + PartialOrd<Self> + Eq + PartialEq<Self> + Hash + Sealed {
    type Key: ParseableKey<Sha256 = Hash, Hash256 = Hash, Ripemd160 = Hash, Hash160 = Hash> + MiniscriptKey;

    fn check_terminal_non_malleable<Pk>(
        _frag: &Terminal<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey;
    fn max_satisfaction_size<Pk>(ms: &Miniscript<Pk, Self>) -> Option<usize>
    where
        Pk: MiniscriptKey;
    fn sig_type() -> SigType;
    fn pk_len<Pk>(pk: &Pk) -> usize
    where
        Pk: MiniscriptKey;
    fn name_str() -> &'static str;

    fn check_witness<Pk>(
        _witness: &[Vec<u8, Global>]
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_global_consensus_validity<Pk>(
        _ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_global_policy_validity<Pk>(
        _ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_local_consensus_validity<Pk>(
        _ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_local_policy_validity<Pk>(
        _ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_global_validity<Pk>(
        ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn check_local_validity<Pk>(
        ms: &Miniscript<Pk, Self>
    ) -> Result<(), ScriptContextError>
    where
        Pk: MiniscriptKey,
    { ... }
    fn top_level_type_check<Pk>(ms: &Miniscript<Pk, Self>) -> Result<(), Error>
    where
        Pk: MiniscriptKey,
    { ... }
    fn other_top_level_checks<Pk>(
        _ms: &Miniscript<Pk, Self>
    ) -> Result<(), Error>
    where
        Pk: MiniscriptKey,
    { ... }
    fn top_level_checks<Pk>(ms: &Miniscript<Pk, Self>) -> Result<(), Error>
    where
        Pk: MiniscriptKey,
    { ... }

}

The ScriptContext for Miniscript. Additional type information associated with miniscript that is used for carrying out checks that dependent on the context under which the script is used. For example, disallowing uncompressed keys in Segwit context

Required Associated Types

type Key: ParseableKey<Sha256 = Hash, Hash256 = Hash, Ripemd160 = Hash, Hash160 = Hash> + MiniscriptKey

The consensus key associated with the type. Must be a parseable key

Required Methods

fn check_terminal_non_malleable<Pk>(
_frag: &Terminal<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Depending on ScriptContext, fragments can be malleable. For Example, under Legacy context, PkH is malleable because it is possible to estimate the cost of satisfaction because of compressed keys This is currently only used in compiler code for removing malleable compilations. This does NOT recursively check if the children of the fragment are valid or not. Since the compilation proceeds in a leaf to root fashion, a recursive check is unnecessary.

fn max_satisfaction_size<Pk>(ms: &Miniscript<Pk, Self>) -> Option<usize>where
Pk: MiniscriptKey,

Depending on script context, the size of a satifaction witness may slightly differ.

fn sig_type() -> SigType

The type of signature required for satisfaction

fn pk_len<Pk>(pk: &Pk) -> usizewhere
Pk: MiniscriptKey,

Get the len of public key when serialized based on context Note that this includes the serialization prefix. Returns 34/66 for Bare/Legacy based on key compressedness 34 for Segwitv0, 33 for Tap

fn name_str() -> &'static str

Local helper function to display error messages with context

Provided Methods

fn check_witness<Pk>(
_witness: &[Vec<u8, Global>]
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Check whether the given satisfaction is valid under the ScriptContext For example, segwit satisfactions may fail if the witness len is more 3600 or number of stack elements are more than 100.

fn check_global_consensus_validity<Pk>(
_ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Depending on script Context, some of the Terminals might not be valid under the current consensus rules. Or some of the script resource limits may have been exceeded. These miniscripts would never be accepted by the Bitcoin network and hence it is safe to discard them For example, in Segwit Context with MiniscriptKey as bitcoin::PublicKey uncompressed public keys are non-standard and thus invalid. In LegacyP2SH context, scripts above 520 bytes are invalid. Post Tapscript upgrade, this would have to consider other nodes. This does NOT recursively check the miniscript fragments.

fn check_global_policy_validity<Pk>(
_ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Depending on script Context, some of the script resource limits may have been exceeded under the current bitcoin core policy rules These miniscripts would never be accepted by the Bitcoin network and hence it is safe to discard them. (unless explicitly disabled by non-standard flag) For example, in Segwit Context with MiniscriptKey as bitcoin::PublicKey scripts over 3600 bytes are invalid. Post Tapscript upgrade, this would have to consider other nodes. This does NOT recursively check the miniscript fragments.

fn check_local_consensus_validity<Pk>(
_ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Consensus rules at the Miniscript satisfaction time. It is possible that some paths of miniscript may exceed resource limits and our current satisfier and lifting analysis would not work correctly. For example, satisfaction path(Legacy/Segwitv0) may require more than 201 opcodes.

fn check_local_policy_validity<Pk>(
_ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Policy rules at the Miniscript satisfaction time. It is possible that some paths of miniscript may exceed resource limits and our current satisfier and lifting analysis would not work correctly. For example, satisfaction path in Legacy context scriptSig more than 1650 bytes

fn check_global_validity<Pk>(
ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Check the consensus + policy(if not disabled) rules that are not based satisfaction

fn check_local_validity<Pk>(
ms: &Miniscript<Pk, Self>
) -> Result<(), ScriptContextError>where
Pk: MiniscriptKey,

Check the consensus + policy(if not disabled) rules including the ones for satisfaction

fn top_level_type_check<Pk>(ms: &Miniscript<Pk, Self>) -> Result<(), Error>where
Pk: MiniscriptKey,

Check whether the top-level is type B

fn other_top_level_checks<Pk>(_ms: &Miniscript<Pk, Self>) -> Result<(), Error>where
Pk: MiniscriptKey,

Other top level checks that are context specific

fn top_level_checks<Pk>(ms: &Miniscript<Pk, Self>) -> Result<(), Error>where
Pk: MiniscriptKey,

Check top level consensus rules.