Module bdk::keys

Key formats

Modules

bip39	keys-bip39
BIP-0039

Structs

DescriptorSinglePriv
A Single Descriptor Secret Key with optional origin information

DescriptorSinglePub
A Single Descriptor Key with optional origin information

GeneratedKey
Output of a GeneratableKey key generation

PrivateKeyGenerateOptions
Options for generating a [PrivateKey]

SortedMultiVec
Contents of a "sortedmulti" descriptor

Enums

DescriptorKey
Container for public or secret keys

DescriptorPublicKey
The MiniscriptKey corresponding to Descriptors. This can either be Single public key or a Xpub

DescriptorSecretKey
A Secret Key that can be either a single key or an Xprv

KeyError
Errors thrown while working with keys

ScriptContextEnum
Enum representation of the known valid ScriptContexts

Traits

DerivableKey
Trait for keys that can be derived.

ExtScriptContext
Trait that adds extra useful methods to ScriptContexts

GeneratableDefaultOptions
Trait that allows generating a key with the default options

GeneratableKey
Trait for keys that can be generated

ScriptContext
The ScriptContext for Miniscript. Additional type information associated with miniscript that is used for carrying out checks that dependent on the context under which the script is used. For example, disallowing uncompressed keys in Segwit context

ToDescriptorKey
Trait for objects that can be turned into a public or secret DescriptorKey

Functions

any_network
Create a set containing mainnet, testnet and regtest

mainnet_network
Create a set only containing mainnet

merge_networks
Compute the intersection of two sets

test_networks
Create a set containing testnet and regtest

Type Definitions

ValidNetworks
Set of valid networks for a key