Struct bdk::keys::bip39::Seed

pub struct Seed { /* fields omitted */ }

This is supported on crate feature keys-bip39 only.

The secret value used to derive HD wallet addresses from a Mnemonic phrase.

Because it is not possible to create a Mnemonic instance that is invalid, it is therefore impossible to have a Seed instance that is invalid. This guarantees that only a valid, intact mnemonic phrase can be used to derive HD wallet addresses.

To get the raw byte value use Seed::as_bytes(). These can be used to derive HD wallet addresses using another crate (deriving HD wallet addresses is outside the scope of this crate and the BIP39 standard).

Seed implements Zeroize, so it’s bytes will be zeroed when it’s dropped.

Implementations

impl Seed

pub fn new(mnemonic: &Mnemonic, password: &str) -> Seed

Generates the seed from the Mnemonic and the password.

pub fn as_bytes(&self) -> &[u8]

Get the seed value as a byte slice