Trait bdk::keys::DerivableKey

pub trait DerivableKey<Ctx: ScriptContext> {
    pub fn add_metadata(
        self, 
        origin: Option<KeySource>, 
        derivation_path: DerivationPath
    ) -> Result<DescriptorKey<Ctx>, KeyError>;
}

Trait for keys that can be derived.

When extra metadata are provided, a DerivableKey can be transofrmed into a DescriptorKey: the trait ToDescriptorKey is automatically implemented for (DerivableKey, DerivationPath) and (DerivableKey, KeySource, DerivationPath) tuples.

For key types that don't encode any indication about the path to use (like bip39), it's generally recommended to implemented this trait instead of ToDescriptorKey. The same rules regarding script context and valid networks apply.

Required methods

pub fn add_metadata(
self,
origin: Option<KeySource>,
derivation_path: DerivationPath
) -> Result<DescriptorKey<Ctx>, KeyError>

Add a extra metadata, consume self and turn it into a DescriptorKey

Loading content...