Trait bdk::keys::GeneratableDefaultOptions

pub trait GeneratableDefaultOptions<Ctx>: GeneratableKey<Ctx> where
    Ctx: ScriptContext,
    Self::Options: Default, {
    fn generate_with_entropy_default(
        entropy: Self::Entropy
    ) -> Result<GeneratedKey<Self, Ctx>, Self::Error> { ... }

    fn generate_default() -> Result<GeneratedKey<Self, Ctx>, Self::Error> { ... }
}

Trait that allows generating a key with the default options

This trait is automatically implemented if the GeneratableKey::Options implements Default.

Provided methods

fn generate_with_entropy_default(
entropy: Self::Entropy
) -> Result<GeneratedKey<Self, Ctx>, Self::Error>

Generate a key with the default options and a given entropy

fn generate_default() -> Result<GeneratedKey<Self, Ctx>, Self::Error>

Generate a key with the default options and a random entropy

Loading content...