Module bdk::wallet

Wallet

This module defines the Wallet structure.

Modules

address_validator

Address validation callbacks

coin_selection

Coin selection

export

Wallet export

signer

Generalized signers

time

Cross-platform time

tx_builder

Transaction builder

Structs

AddressInfo

A derived address and the index it was found at For convenience this automatically derefs to Address

Wallet

A Bitcoin wallet

Enums

AddressIndex

The address index selection strategy to use to derived an address from the wallet’s external descriptor. See Wallet::get_address. If you’re unsure which one to use use WalletIndex::New.

Traits

IsDust

Trait to check if a value is below the dust limit

Vbytes

Trait implemented by types that can be used to measure weight units.