Struct bdk::wallet::tx_builder::TxBuilder

pub struct TxBuilder<D: Database, Cs: CoinSelectionAlgorithm<D>, Ctx: TxBuilderContext> { /* fields omitted */ }

A transaction builder

This structure contains the configuration that the wallet must follow to build a transaction.

For an example see this module's documentation;

Implementations

impl<D: Database, Ctx: TxBuilderContext> TxBuilder<D, DefaultCoinSelectionAlgorithm, Ctx>

pub fn new() -> Self

Create an empty builder

impl<D: Database, Cs: CoinSelectionAlgorithm<D>, Ctx: TxBuilderContext> TxBuilder<D, Cs, Ctx>

pub fn fee_rate(self, fee_rate: FeeRate) -> Self

Set a custom fee rate

pub fn fee_absolute(self, fee_amount: u64) -> Self

Set an absolute fee

pub fn policy_path(
self,
policy_path: BTreeMap<String, Vec<usize>>,
keychain: KeychainKind
) -> Self

Set the policy path to use while creating the transaction for a given keychain.

This method accepts a map where the key is the policy node id (see Policy::id) and the value is the list of the indexes of the items that are intended to be satisfied from the policy node (see SatisfiableItem::Thresh::items).

Example

An example of when the policy path is needed is the following descriptor: wsh(thresh(2,pk(A),sj:and_v(v:pk(B),n:older(6)),snj:and_v(v:pk(C),after(630000)))), derived from the miniscript policy thresh(2,pk(A),and(pk(B),older(6)),and(pk(C),after(630000))). It declares three descriptor fragments, and at the top level it uses thresh() to ensure that at least two of them are satisfied. The individual fragments are:

pk(A)

and(pk(B),older(6))

and(pk(C),after(630000))

When those conditions are combined in pairs, it's clear that the transaction needs to be created differently depending on how the user intends to satisfy the policy afterwards:

If fragments 1 and 2 are used, the transaction will need to use a specific n_sequence in order to spend an OP_CSV branch.

If fragments 1 and 3 are used, the transaction will need to use a specific locktime in order to spend an OP_CLTV branch.

If fragments 2 and 3 are used, the transaction will need both.

When the spending policy is represented as a tree (see Wallet::policies), every node is assigned a unique identifier that can be used in the policy path to specify which of the node's children the user intends to satisfy: for instance, assuming the thresh() root node of this example has an id of aabbccdd, the policy path map would look like:

{ "aabbccdd" => [0, 1] }

where the key is the node's id, and the value is a list of the children that should be used, in no particular order.

If a particularly complex descriptor has multiple ambiguous thresholds in its structure, multiple entries can be added to the map, one for each node that requires an explicit path.

let mut path = BTreeMap::new();
path.insert("aabbccdd".to_string(), vec![0, 1]);

let builder = TxBuilder::with_recipients(vec![(to_address.script_pubkey(), 50_000)])
    .policy_path(path, KeychainKind::External);

pub fn utxos(self, utxos: Vec<OutPoint>) -> Self

Replace the internal list of utxos that must be spent with a new list

These have priority over the "unspendable" utxos, meaning that if a utxo is present both in the "utxos" and the "unspendable" list, it will be spent.

pub fn add_utxo(self, utxo: OutPoint) -> Self

Add a utxo to the internal list of utxos that must be spent

These have priority over the "unspendable" utxos, meaning that if a utxo is present both in the "utxos" and the "unspendable" list, it will be spent.

pub fn manually_selected_only(self) -> Self

Only spend utxos added by add_utxo and utxos.

The wallet will not add additional utxos to the transaction even if they are needed to make the transaction valid.

pub fn unspendable(self, unspendable: Vec<OutPoint>) -> Self

Replace the internal list of unspendable utxos with a new list

It's important to note that the "must-be-spent" utxos added with TxBuilder::utxos and TxBuilder::add_utxo have priority over these. See the docs of the two linked methods for more details.

pub fn add_unspendable(self, unspendable: OutPoint) -> Self

Add a utxo to the internal list of unspendable utxos

It's important to note that the "must-be-spent" utxos added with TxBuilder::utxos and TxBuilder::add_utxo have priority over this. See the docs of the two linked methods for more details.

pub fn sighash(self, sighash: SigHashType) -> Self

Sign with a specific sig hash

Use this option very carefully

pub fn ordering(self, ordering: TxOrdering) -> Self

Choose the ordering for inputs and outputs of the transaction

pub fn nlocktime(self, locktime: u32) -> Self

Use a specific nLockTime while creating the transaction

This can cause conflicts if the wallet's descriptors contain an "after" (OP_CLTV) operator.

pub fn version(self, version: i32) -> Self

Build a transaction with a specific version

The version should always be greater than 0 and greater than 1 if the wallet's descriptors contain an "older" (OP_CSV) operator.

pub fn do_not_spend_change(self) -> Self

Do not spend change outputs

This effectively adds all the change outputs to the "unspendable" list. See TxBuilder::unspendable.

pub fn only_spend_change(self) -> Self

Only spend change outputs

This effectively adds all the non-change outputs to the "unspendable" list. See TxBuilder::unspendable.

pub fn change_policy(self, change_policy: ChangeSpendPolicy) -> Self

Set a specific ChangeSpendPolicy. See TxBuilder::do_not_spend_change and TxBuilder::only_spend_change for some shortcuts.

pub fn force_non_witness_utxo(self) -> Self

Fill-in the psbt::Input::non_witness_utxo field even if the wallet only has SegWit descriptors.

This is useful for signers which always require it, like Trezor hardware wallets.

pub fn include_output_redeem_witness_script(self) -> Self

Fill-in the psbt::Output::redeem_script and psbt::Output::witness_script fields.

This is useful for signers which always require it, like ColdCard hardware wallets.

pub fn add_global_xpubs(self) -> Self

Fill-in the PSBT_GLOBAL_XPUB field with the extended keys contained in both the external and internal descriptors

This is useful for offline signers that take part to a multisig. Some hardware wallets like BitBox and ColdCard are known to require this.

pub fn drain_wallet(self) -> Self

Spend all the available inputs. This respects filters like TxBuilder::unspendable and the change policy.

pub fn coin_selection<P: CoinSelectionAlgorithm<D>>(
self,
coin_selection: P
) -> TxBuilder<D, P, Ctx>

Choose the coin selection algorithm

Overrides the DefaultCoinSelectionAlgorithm.

impl<D: Database> TxBuilder<D, DefaultCoinSelectionAlgorithm, CreateTx>

pub fn with_recipients(recipients: Vec<(Script, u64)>) -> Self

Create a builder starting from a list of recipients

impl<D: Database, Cs: CoinSelectionAlgorithm<D>> TxBuilder<D, Cs, CreateTx>

pub fn set_recipients(self, recipients: Vec<(Script, u64)>) -> Self

Replace the recipients already added with a new list

pub fn add_recipient(self, script_pubkey: Script, amount: u64) -> Self

Add a recipient to the internal list

pub fn set_single_recipient(self, recipient: Script) -> Self

Set a single recipient that will get all the selected funds minus the fee. No change will be created

This method overrides any recipient set with set_recipients or add_recipient.

It can only be used in conjunction with drain_wallet to send the entire content of the wallet (minus filters) to a single recipient or with a list of manually selected UTXOs by enabling manually_selected_only and selecting them with utxos or add_utxo.

When bumping the fees of a transaction made with this option, the user should remeber to add maintain_single_recipient to correctly update the single output instead of adding one more for the change.

pub fn enable_rbf(self) -> Self

Enable signaling RBF

This will use the default nSequence value of 0xFFFFFFFD.

pub fn enable_rbf_with_sequence(self, nsequence: u32) -> Self

Enable signaling RBF with a specific nSequence value

This can cause conflicts if the wallet's descriptors contain an "older" (OP_CSV) operator and the given nsequence is lower than the CSV value.

If the nsequence is higher than 0xFFFFFFFD an error will be thrown, since it would not be a valid nSequence to signal RBF.

impl<D: Database> TxBuilder<D, DefaultCoinSelectionAlgorithm, BumpFee>

pub fn maintain_single_recipient(self) -> Self

Bump the fees of a transaction made with set_single_recipient

Unless extra inputs are specified with add_utxo or utxos, this flag will make bump_fee reduce the value of the existing output, or fail if it would be consumed entirely given the higher new fee rate.

If extra inputs are added and they are not entirely consumed in fees, a change output will not be added; the existing output will simply grow in value.

Fails if the transaction has more than one outputs.